requirements-production.txt
# ===========================================
# PRODUCTION REQUIREMENTS
# ===========================================
# Production-ready Flask application with security dependencies

# Core Flask and Web Framework
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7

# Database
SQLite3  # Built into Python

# Security Dependencies
cryptography==41.0.7          # For data encryption and secure key derivation
python-dotenv==1.0.0          # For environment variable management
bleach==6.1.0                 # For HTML sanitization and XSS prevention
python-magic==0.4.27          # For file type validation
python-magic-bin==0.4.14      # Windows binary for python-magic

# Password and Authentication Security
passlib[bcrypt]==1.7.4        # Advanced password hashing
argon2-cffi==23.1.0           # Argon2 password hashing (recommended)

# Rate Limiting and Request Security
Flask-Limiter==3.5.0          # Rate limiting for Flask
redis==5.0.1                  # For distributed rate limiting storage

# Session Security
Flask-Session==0.5.0          # Server-side session management

# File Processing and Document Handling
PyPDF2==3.0.1                 # PDF processing
python-docx==0.8.11           # Word document processing

# Email Security
email-validator==2.1.0        # Email validation

# Logging and Monitoring
python-json-logger==2.0.7     # JSON structured logging

# WSGI Server (Production)
gunicorn==21.2.0              # Production WSGI server
gevent==23.9.1                # Async worker for gunicorn

# SSL/TLS Security
pyOpenSSL==23.3.0             # SSL certificate handling
certifi==2023.7.22            # CA certificate bundle

# Content Security
Pillow==10.1.0                # Image processing and validation
defusedxml==0.7.1             # XML bomb protection

# Development and Testing (optional)
pytest==7.4.3                # Testing framework
pytest-flask==1.3.0          # Flask testing utilities
coverage==7.3.2              # Code coverage

# Database Backup and Encryption
cryptg==0.4.0                # Additional cryptography support

# System and OS Security
psutil==5.9.6                # System monitoring

# HTTP Security Headers
flask-talisman==1.1.0        # Security headers middleware

# Input Validation
marshmallow==3.20.1          # Schema validation
jsonschema==4.19.2           # JSON schema validation

# Additional Security Tools
secrets                       # Built into Python 3.6+
hashlib                       # Built into Python
hmac                          # Built into Python
base64                        # Built into Python
